ready for testing on QA server
fix bug: relevant blah should be sorted by cohort-strength from strong to weak, not weak to strong
fix bug: unsafe inbox number string is "L" and "F", not "LS" and "FS"

+++ src/main/java/com/eweware/service/base/CommonUtilities.java
@@ -352,7 +352,7 @@ public final class CommonUtilities {
      * @param inboxNumber The inbox number
      * @return  Returns the collection name of the inbox
      */
-    public static String makeInboxCollectionName(String groupId, String cohortId, Integer inboxNumber, Boolean safe) {
+    public static String makeInboxCollectionName(String groupId, String cohortId, Long inboxNumber, Boolean safe) {
         final StringBuilder inboxCollectionName = new StringBuilder(groupId);
         inboxCollectionName.append("-");
         inboxCollectionName.append(cohortId);
